requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
pdfplumber==0.11.0
python-docx==1.1.0
